Quick fix for performance issues with long assignments
Long assignments with line based grading caused errors in Underscore:
"Max Call Stack Size Exceeded". Since the code only wanted the length
of the longest array, there's no need to run it through map and max
when a normal for loop works better and faster. This seems to get rid
of the error.

The grading is still quite inefficient for solutions with a lot of
lines, but this should be enough for an immediate fix.